+++ docs/history.rst
@@ -5,11 +5,17 @@ History
 =======
 
 
-Version 0.4.3 - Unreleased
+Version 0.4.4 - Unreleased
 --------------------------
 
 This is the last Alpha branch version, it have everything working still needs some
 improvements and minor features to come with Beta branch versions.
+
+* Improved image crafting for test and demo usage with new classes
+  ``SampleImageCrafter`` and ``DjangoSampleImageCrafter``;
+
+Version 0.4.3 - Unreleased
+--------------------------
 
 * Article, Category and Author models;
 * Article and Category translations;